Commented out svg drawing temporarily - it is too slow Moved the SVG objects and in general the NIPS environment to NIPSEnvironment to keep a clean distinction Modified dynamics to account for angular velocities, collisions! Added some synchronicity Added the ability for old drops to be thrown out (DiscoInterfaceClientHandler)

+++ sim/src/org/rlcommunity/critter/Wall.java
@@ -78,7 +78,7 @@ public class Wall extends SimulatorObject {
 
 		double points[] = new double[walls.size()];
 		for (int i = 0; i < points.length; i++) {
-			points[i] = (int) walls.get(i).y;
+			points[i] = walls.get(i).y;
 		}
 		return points;
 	}
